# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/v1/employees": {
      "post": {
        "operationId": "EmployeesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new employee",
        "tags": [
          "Employees"
        ]
      },
      "get": {
        "operationId": "EmployeesController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by role",
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "MANAGER",
                "RECEPTIONIST",
                "HOUSEKEEPING"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "ON_LEAVE",
                "TERMINATED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of employees"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all employees with pagination and filters",
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee details"
          },
          "404": {
            "description": "Employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get employee by ID",
        "tags": [
          "Employees"
        ]
      },
      "patch": {
        "operationId": "EmployeesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update employee",
        "tags": [
          "Employees"
        ]
      },
      "delete": {
        "operationId": "EmployeesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete employee (soft delete)",
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/v1/employees/{id}/change-password": {
      "patch": {
        "operationId": "EmployeesController_changePasswordAdmin_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change employee password (Admin only)",
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/v1/employees/me/change-password": {
      "patch": {
        "operationId": "EmployeesController_changeOwnPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Current password is incorrect"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change own password",
        "tags": [
          "Employees"
        ]
      }
    },
    "/api/v1/v1/employees/{id}/restore": {
      "post": {
        "operationId": "EmployeesController_restore_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Employee UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee restored successfully"
          },
          "404": {
            "description": "Deleted employee not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore deleted employee",
        "tags": [
          "Employees"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotel Management API",
    "description": "API for hotel management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@hotel.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "SecurePass123!",
            "description": "Password"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "RECEPTIONIST",
              "HOUSEKEEPING"
            ],
            "default": "RECEPTIONIST",
            "description": "Employee role"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "ON_LEAVE",
              "TERMINATED"
            ],
            "default": "ACTIVE",
            "description": "Employee status"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "RECEPTIONIST",
              "HOUSEKEEPING"
            ],
            "default": "RECEPTIONIST",
            "description": "Employee role"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "ON_LEAVE",
              "TERMINATED"
            ],
            "default": "ACTIVE",
            "description": "Employee status"
          },
          "password": {
            "type": "string",
            "example": "NewSecurePass123!",
            "description": "New password"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "example": "NewSecurePass123!",
            "description": "New password"
          }
        }
      },
...
    }
  }
}